# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponseObject": {
        "properties": {
          "data": {},
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "message",
          "success"
        ],
        "type": "object"
      },
...
      "DictionaryItemEdit": {
        "properties": {
          "color": {
            "pattern": "default|success|warning|error|processing|blue|purple|cyan|green|orange|red",
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "label": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "sortOrder": {
            "format": "int32",
            "maximum": 9999,
            "minimum": 0,
            "type": "integer"
          },
          "value": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "label",
          "value"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "通过滑块验证与登录接口取得的一次性会话令牌",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "统一响应、动作权限、数据范围与乐观锁均由服务端执行。关闭模块不开放接口。",
    "title": "Mayday API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/system/dictionaries/{typeId}/items": {
      "get": {
        "operationId": "DictionaryController_list",
        "parameters": [
          {
            "in": "path",
            "name": "typeId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "enabled",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 10,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "dictionary-controller"
        ]
      },
      "post": {
        "operationId": "DictionaryController_create",
        "parameters": [
          {
            "in": "path",
            "name": "typeId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryItemEdit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "dictionary-controller"
        ]
      }
    },
    "/api/system/dictionaries/{typeId}/items/{id}": {
      "delete": {
        "operationId": "DictionaryController_delete",
        "parameters": [
          {
            "in": "path",
            "name": "typeId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "dictionary-controller"
        ]
      },
      "put": {
        "operationId": "DictionaryController_update",
        "parameters": [
          {
            "in": "path",
            "name": "typeId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryItemEdit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "dictionary-controller"
        ]
      }
    },
    "/api/system/dictionary-options/{code}": {
      "get": {
        "operationId": "DictionaryController_options",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "dictionary-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}